# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de gastos e informações financeiras",
        "title": "Contabius API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://vkunssec.vercel.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/vkunssec/contabius/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "${HOST}",
    "basePath": "/",
    "paths": {
...
        "/investment/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar um investimento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Rota para atualizar um investimento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do investimento",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Investments"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID do investimento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/investments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar os investimentos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Rota para retornar os investimentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5,
                                6,
                                7,
                                8,
                                9,
                                10,
                                11,
                                12,
                                13,
                                14,
                                15
                            ],
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "description": "Ids dos investimentos a serem retornados",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Investments"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para criar um investimento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Rota para criar um investimento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do investimento",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InvestmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/investments/get_types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar todos os tipos de investimentos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Rota para retornar todos os tipos de investimentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5,
                                6,
                                7,
                                8,
                                9,
                                10,
                                11,
                                12,
                                13,
                                14,
                                15
                            ],
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "description": "Ids dos investimentos a serem retornados",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetInvestmentsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/revenues/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para deletar uma receita",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Revenues"
                ],
                "summary": "Rota para deletar uma receita",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da receita",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Accounts": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "Nome da conta",
                    "type": "string",
                    "example": "Conta Corrente"
                },
                "color": {
                    "description": "Cor da conta",
                    "type": "string",
                    "example": "#000000"
                },
                "created_at": {
                    "description": "Data de criação",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID da conta",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "updated_at": {
                    "description": "Data de atualização",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "BadRequest": {
            "description": "Estrutura padrão de resposta de erro de requisição",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Erro ao processar a requisição"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "Currency": {
            "type": "string",
            "enum": [
                "BRL",
                "USD",
                "EUR"
            ],
            "x-enum-varnames": [
                "CurrencyBRL",
                "CurrencyUSD",
                "CurrencyEUR"
            ]
        },
...
        "GetInvestmentsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem de sucesso ou erro",
                    "type": "string",
                    "example": "Investimentos retornados com sucesso"
                },
                "success": {
                    "description": "Sucesso da operação",
                    "type": "boolean",
                    "example": true
                },
                "types": {
                    "description": "Dados dos investimentos",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InvestmentType"
                    }
                }
            }
        },
...
        "InternalServerError": {
            "description": "Estrutura padrão de resposta de erro de servidor interno",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "InvestmentId": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "InvestmentTypeCDIId",
                "InvestmentTypeCDBId",
                "InvestmentTypeLCAId",
                "InvestmentTypeLCIId",
                "InvestmentTypePoupancaId",
                "InvestmentTypeTesouroDiretoId",
                "InvestmentTypeFundosInvestimentoId",
                "InvestmentTypeAcoesId",
                "InvestmentTypeFIIsId",
                "InvestmentTypeCRIId",
                "InvestmentTypeCRAId",
                "InvestmentTypeDebenturesId",
                "InvestmentTypeBDRId",
                "InvestmentTypeCOEId",
                "InvestmentTypeOtherId"
            ]
        },
        "InvestmentLiteral": {
            "type": "string",
            "enum": [
                "cdi",
                "cdb",
                "lca",
                "lci",
                "poupanca",
                "tesouro_direto",
                "fundos_investimento",
                "acoes",
                "fiis",
                "cri",
                "cra",
                "debentures",
                "bdr",
                "coe",
                "other"
            ],
            "x-enum-comments": {
                "InvestmentTypeAcoes": "Ações",
                "InvestmentTypeBDR": "Depositário Brasileiro de Recebíveis",
                "InvestmentTypeCDB": "Certificado de Depósito Bancário",
                "InvestmentTypeCDI": "Certificado de Depósito Interbancário",
                "InvestmentTypeCOE": "Certificado de Operações Estruturadas",
                "InvestmentTypeCRA": "Certificado de Recebíveis do Agronegócio",
                "InvestmentTypeCRI": "Certificado de Recebíveis Imobiliários",
                "InvestmentTypeDebentures": "Debêntures",
                "InvestmentTypeFIIs": "Fundos de Investimento Imobiliário",
                "InvestmentTypeFundosInvestimento": "Fundos de Investimento",
                "InvestmentTypeLCA": "Letra de Crédito do Agronegócio",
                "InvestmentTypeLCI": "Letra de Crédito Imobiliário",
                "InvestmentTypeOther": "Outros",
                "InvestmentTypePoupanca": "Poupança",
                "InvestmentTypeTesouroDireto": "Tesouro Direto"
            },
            "x-enum-varnames": [
                "InvestmentTypeCDI",
                "InvestmentTypeCDB",
                "InvestmentTypeLCA",
                "InvestmentTypeLCI",
                "InvestmentTypePoupanca",
                "InvestmentTypeTesouroDireto",
                "InvestmentTypeFundosInvestimento",
                "InvestmentTypeAcoes",
                "InvestmentTypeFIIs",
                "InvestmentTypeCRI",
                "InvestmentTypeCRA",
                "InvestmentTypeDebentures",
                "InvestmentTypeBDR",
                "InvestmentTypeCOE",
                "InvestmentTypeOther"
            ]
        },
        "InvestmentRequest": {
            "type": "object",
            "required": [
                "account",
                "amount",
                "investment_type"
            ],
            "properties": {
                "account": {
                    "description": "Conta do usuário",
                    "allOf": [
                        {
                            "$ref": "#/definitions/PartialAccountRequest"
                        }
                    ]
                },
                "amount": {
                    "description": "Valor do investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "description": {
                    "description": "Descrição do investimento",
                    "type": "string",
                    "example": "Investimento em CDB"
                },
                "investment_type": {
                    "description": "Tipo de investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/InvestmentType"
                        }
                    ]
                },
                "recurrence": {
                    "description": "Recurrence do investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Recurrence"
                        }
                    ]
                },
                "recurrence_day": {
                    "description": "Dia da recorrência",
                    "type": "integer"
                }
            }
        },
        "InvestmentType": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID do investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/InvestmentId"
                        }
                    ],
                    "example": 1
                },
                "investment": {
                    "description": "Tipo de investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/InvestmentLiteral"
                        }
                    ],
                    "example": "cdi"
                }
            }
        },
        "Investments": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "Conta do usuário",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Accounts"
                        }
                    ]
                },
                "amount": {
                    "description": "Valor do investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "created_at": {
                    "description": "Data de criação do investimento",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "description": {
                    "description": "Descrição do investimento",
                    "type": "string",
                    "example": "Investimento em CDB"
                },
                "id": {
                    "description": "ID do investimento",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "investment_type": {
                    "description": "Tipo de investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/InvestmentType"
                        }
                    ]
                },
                "recurrence": {
                    "description": "Recurrence do investimento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Recurrence"
                        }
                    ],
                    "example": "monthly"
                },
                "recurrence_day": {
                    "description": "Dia da recorrência",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "Data de atualização do investimento",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
...
        "Money": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "default code BRL",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Currency"
                        }
                    ],
                    "example": "BRL"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
        "PartialAccountRequest": {
            "type": "object",
            "required": [
                "account",
                "color",
                "id"
            ],
            "properties": {
                "account": {
                    "description": "Nome da conta",
                    "type": "string",
                    "example": "Conta Corrente"
                },
                "color": {
                    "description": "Cor da conta",
                    "type": "string",
                    "example": "#000000"
                },
                "id": {
                    "description": "ID da conta",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                }
            }
        },
...
        "Recurrence": {
            "type": "string",
            "enum": [
                "monthly",
                "yearly",
                "sporadic"
            ],
            "x-enum-comments": {
                "RecurrenceMonthly": "Mensal",
                "RecurrenceSporadic": "Esporádico",
                "RecurrenceYearly": "Anual"
            },
            "x-enum-varnames": [
                "RecurrenceMonthly",
                "RecurrenceYearly",
                "RecurrenceSporadic"
            ]
        },
        "Response": {
            "description": "Estrutura padrão de resposta",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Dados da resposta"
                },
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}